Move locales.ts from /src/utils folder to new folder /src/l10n.
