+++ src/blocks/FooterBlock/FooterBlock.stories.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 // import StoryRouter from 'storybook-react-router'
 import { Provider } from 'constate'
 import { globalStoryMock, minimalColorMap } from 'utils/storybook'
-import { fallbackLocale } from 'utils/locales'
+import { fallbackLocale } from 'src/l10n/locales'
 import { Footer } from './FooterBlock'
 
 export default {
