+++ src/components/Perils/Perils.stories.tsx
@@ -1,7 +1,7 @@
 import { select, withKnobs } from '@storybook/addon-knobs'
 import React from 'react'
 import { globalStoryMock } from 'utils/storybook'
-import { locales, fallbackLocale } from 'utils/locales'
+import { locales, fallbackLocale } from 'src/l10n/locales'
 import { TypeOfContract } from './types'
 import { Perils } from './'
