+++ src/l10n/locales.ts
