+++ src/test-utils.tsx
@@ -4,7 +4,7 @@ import { Provider } from 'constate'
 import { HelmetProvider } from 'react-helmet-async'
 import { v4 as uuidV4 } from 'uuid'
 import { BrowserRouter } from 'react-router-dom'
-import { locales } from 'utils/locales'
+import { locales } from 'src/l10n/locales'
 import { LocaleProvider } from './context/LocaleContext'
 
 const initialState = {
